Created DamageDealer and Enemy scripts
DamageDealer: This is created for the different lasers to apply damage to whichever object they hit (player or enemy) and to destroy themselves on hit.

Enemy: Created so that damage can be applied to the health if hit by a player laser and enemies are set to also shoot lasers at random intervals using a counter.